[REFACTOR] Move functions to isochrone.h
